refactor del frament home  -> home_menu
