+++ .idea/misc.xml
@@ -7,6 +7,9 @@
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/activity_main.xml" value="0.33" />
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/adapter_menu_list.xml" value="0.24947916666666667" />
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/fragment_home.xml" value="0.6084860173577628" />
+        <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/nav_header_landing_menu.xml" value="0.31354166666666666" />
+        <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/menu/activity_landing_menu_drawer.xml" value="0.31354166666666666" />
+        <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/menu/landing_menu.xml" value="0.33" />
       </map>
     </option>
   </component>
