+++ app/src/main/res/layout/nav_header_landing_menu.xml
@@ -10,7 +10,11 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    android:theme="@style/ThemeOverlay.AppCompat.Dark">
+    android:theme="@style/ThemeOverlay.AppCompat.Dark"
+
+    android:layout_marginTop="50dp"
+    >
+<!--    valor agregado layout margin top -->
 
     <ImageView
         android:id="@+id/imageView"
