+++ app/src/main/res/menu/activity_landing_menu_drawer.xml
@@ -5,9 +5,9 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_home"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="@string/menu_home" />
+            android:id="@+id/nav_menu"
+            android:icon="@drawable/logo_prototype"
+            android:title="@string/menu_home_pizzas_menu" />
         <item
             android:id="@+id/nav_gallery"
             android:icon="@drawable/ic_menu_gallery"
