+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -3,13 +3,13 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/nav_home">
+    app:startDestination="@+id/nav_menu">
 
     <fragment
-        android:id="@+id/nav_home"
-        android:name="sv.edu.udb.pizzeria_forkify.OrderPizza.ui.home.HomeFragment"
-        android:label="@string/menu_home"
-        tools:layout="@layout/fragment_home" />
+        android:id="@+id/nav_menu"
+        android:name="sv.edu.udb.pizzeria_forkify.OrderPizza.ui.home_menu.HomeMenuFragment"
+        android:label="@string/menu_home_pizzas_menu"
+        tools:layout="@layout/fragment_home_menu" />
 
     <fragment
         android:id="@+id/nav_gallery"
